Add change tracker support for read-only properties with backing fields
Issue #4461

Previously the materializer would use the backing field, but then when a value needed to be set for fixup, etc, things would fall down. Now we generate delegates everywhere that can use the backing field.

Note that query still prefers the backing field, while fixup prefers the property.

MemberMapper was made a service of the model. Only the minimal done here--we can allow it to be set by convention or fluent API if the need to customize it is requested.

Also fixed #758 - When mapping to backing fields handle cases where field is on base type